added user authentication & authorization

+++ src/main/java/c3phal0p0d/projects/quizzical/quiz/QuizRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuizRepository extends CrudRepository<Quiz, Integer> {
+public interface QuizRepository extends CrudRepository<Quiz, Long> {
 
 }
